requirements.txt
fastapi
python-fasthtml
uvicorn
requests
python-multipart
python-dotenv
Flask
Flask-Login
Flask-SQLAlchemy
beautifulsoup4
tqdm
sentence-transformers
numpy
supabase
openai
